Refactor: reimplement Address type as a `struct`.
The `Address` type is one of the most heavily used types in Reko. By changing it from `class` to `struct`, we greatly reduce the pressure on the garbage collector. The run time of the Reko unit tests and the regression tests is 30% lower after this change.

+++ src/Core/Output/OutputFilePolicy.cs
@@ -97,9 +97,11 @@ namespace Reko.Core.Output
                 var strGlobals = pt.Pointee.ResolveAs<StructureType>();
                 if (strGlobals != null)
                 {
+                    //$TODO: the cast to (uint) below needs to be changed to (ulong)
+                    // once the StructureField.Offset property is widened to 64 bits.
                     return strGlobals.Fields.Select(field =>
                         (field,
-                         program.Platform.MakeAddressFromLinear((ulong) field.Offset, false)));
+                         program.Platform.MakeAddressFromLinear((uint) field.Offset, false)));
                 }
             }
             return new (StructureField, Address)[0];
